Point 10: Move importer() to DynamicLoader
- Moved functions load_modules(path), importer(usable_modules) and _process_spec(spec) to the only class that used these functions: DynamicLoader. Further renamed functions to prevent duplicated function names.
