+++ src/main/python/utilities/utilities.py
@@ -2,7 +2,6 @@ import os
 import warnings
 import functools
 from pathlib import Path
-from importlib import util
 
 
 def get_project_root_dir():
@@ -12,57 +11,6 @@ def get_project_root_dir():
     """
     #  This script is 5 directories deep into the project.
     return Path(__file__).parent.parent.parent.parent.parent
-
-
-def load_modules(path):
-    """
-    Utilities function to dynamically load in modules in the given path
-    :param path: path to the modules
-    :return: list
-    """
-    modules = []
-    for module in os.listdir(path):
-        module = os.path.join(path, module)
-        if module.endswith('.py') and \
-                not module.endswith('__.py') and \
-                not module.endswith('abstract.py'):
-            modules.append(module)
-    return modules
-
-
-def importer(usable_modules: list):
-    """
-    Utilitarian function for the imputer and preprocessor to dynamically load
-    in the modules using the import_module library.
-    :param usable_modules: list of absolute paths to potential modules
-    :return: list of usable modules
-    """
-    return_modules = {}
-    for module in usable_modules:
-        name = os.path.basename(module).split('.py')[0]
-        spec = util.spec_from_file_location(
-            name=name,
-            location=module
-        )
-        loaded_module = _process_spec(spec)
-        if loaded_module and module not in return_modules.keys():
-            return_modules[module] = loaded_module
-    return return_modules
-
-
-def _process_spec(spec):
-    return_spec = None
-    loaded_spec = util.module_from_spec(spec)
-    spec.loader.exec_module(loaded_spec)
-    for attribute in dir(loaded_spec):
-        if not attribute.startswith('Template') and \
-                not attribute.startswith('__'):
-            get_attribute = getattr(loaded_spec, attribute)
-            if 'name' in dir(get_attribute) and \
-                    'usable' in dir(get_attribute) and \
-                    get_attribute().usable is True:
-                return_spec = get_attribute()
-    return return_spec
 
 
 def get_filename_and_extension(path):
